requirements.txt
pytest
request
pytest-html
pytest-xdist